+++ api/github.ts
@@ -1,5 +1,12 @@
 import { VercelRequest, VercelResponse } from '@vercel/node';
 import axios from 'axios';
+import { 
+  validateGitHubContentsResponse,
+  validateGitHubSearchResponse,
+  safeValidateGitHubContentsResponse,
+  safeValidateGitHubSearchResponse,
+  ApiErrorResponseSchema 
+} from '../src/services/github/schemas/apiSchemas';
 
 // 配置常量
 const GITHUB_API_BASE = 'https://api.github.com';
@@ -226,7 +233,17 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
           })
         );
         
-        return res.status(200).json(response.data);
+        // 验证GitHub API响应结构
+        const validation = safeValidateGitHubContentsResponse(response.data);
+        if (!validation.success) {
+          console.error('GitHub API响应验证失败:', validation.error);
+          return res.status(500).json({ 
+            error: '获取内容失败',
+            message: `API响应格式错误: ${validation.error}`
+          });
+        }
+        
+        return res.status(200).json(validation.data);
       } catch (error: any) {
         console.error('GitHub API请求失败:', error.message);
         
@@ -328,18 +345,36 @@ export default async function handler(req: VercelRequest, res: VercelResponse) {
 
       const searchQuery = `repo:${repoOwner}/${repoName} ${qParam}`;
       
-      const response = await handleRequestWithRetry(() => 
-        axios.get(`${GITHUB_API_BASE}/search/code`, { 
-          headers: getAuthHeaders(),
-          params: {
-            q: searchQuery,
-            sort: (Array.isArray(sort) ? sort[0] : sort) || 'best-match',
-            order: (Array.isArray(order) ? order[0] : order) || 'desc'
-          }
-        })
-      );
-      
-      return res.status(200).json(response.data);
+      try {
+        const response = await handleRequestWithRetry(() => 
+          axios.get(`${GITHUB_API_BASE}/search/code`, { 
+            headers: getAuthHeaders(),
+            params: {
+              q: searchQuery,
+              sort: (Array.isArray(sort) ? sort[0] : sort) || 'best-match',
+              order: (Array.isArray(order) ? order[0] : order) || 'desc'
+            }
+          })
+        );
+        
+        // 验证GitHub搜索API响应结构
+        const validation = safeValidateGitHubSearchResponse(response.data);
+        if (!validation.success) {
+          console.error('GitHub搜索API响应验证失败:', validation.error);
+          return res.status(500).json({ 
+            error: '搜索失败',
+            message: `搜索响应格式错误: ${validation.error}`
+          });
+        }
+        
+        return res.status(200).json(validation.data);
+      } catch (error: any) {
+        console.error('GitHub搜索API请求失败:', error.message);
+        return res.status(error.response?.status || 500).json({ 
+          error: '搜索失败',
+          message: error.message
+        });
+      }
     }
     
     // 未知操作
