+++ package.json
@@ -42,7 +42,6 @@
     "jszip": "^3.10.1",
     "katex": "^0.16.22",
     "notistack": "^3.0.2",
-    
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-helmet-async": "^2.0.5",
@@ -54,7 +53,8 @@
     "rehype-katex": "^7.0.1",
     "rehype-raw": "^7.0.0",
     "remark-gfm": "^4.0.1",
-    "remark-math": "^6.0.0"
+    "remark-math": "^6.0.0",
+    "zod": "^4.1.9"
   },
   "overrides": {
     "react-helmet-async": {
